+++ bigmler/bigmler.py
@@ -68,7 +68,7 @@ from bigmler.options import create_parser
 from bigmler.defaults import get_user_defaults
 from bigmler.defaults import DEFAULTS_FILE
 from bigmler.prediction import predict, combine_votes
-from bigmler.prediction import MAX_MODELS
+from bigmler.prediction import MAX_MODELS, FULL_FORMAT
 from bigmler.train_reader import TrainReader
 
 
@@ -422,7 +422,10 @@ def multi_label_expansion(training_set, training_set_header, objective_field,
                        for label in training_reader.labels]
     new_headers.extend(new_field_names)
     new_headers.append(training_reader.objective_name)
-    file_name = os.path.basename(training_set)
+    try:
+        file_name = os.path.basename(training_set)
+    except AttributeError:
+        file_name = "training_set.csv"
     output_file = "%s%sextended_%s" % (output_path, os.sep, file_name)
     message = u.dated("Transforming to extended source.\n")
     u.log_message(message, log_file=session_file,
@@ -517,7 +520,7 @@ def compute_output(api, args, training_set, test_set=None, output=None,
                 pass
 
     # labels to be used in multi-label expansion
-    labels = (map(lambda x: x.strip(), args.labels.split(','))
+    labels = (map(str.strip, args.labels.split(','))
               if args.labels is not None else None)
     if labels is not None:
         labels = sorted(labels)
@@ -648,14 +651,22 @@ def main(args=sys.argv[1:]):
 
     """
     train_stdin = False
+    test_stdin = False
     flags = []
     for i in range(0, len(args)):
         if args[i].startswith("--"):
-            args[i] = args[i].replace("_", "-")
-            flags.append(args[i])
-            if (args[i] == '--train' and
+            flag = args[i]
+            # syntax --flag=value
+            if "=" in flag:
+                flag = args[i][0: flag.index("=")]
+            flag = flag.replace("_", "-")
+            flags.append(flag)
+            if (flag == '--train' and
                     (i == len(args) - 1 or args[i + 1].startswith("--"))):
                 train_stdin = True
+            elif (flag == '--test' and
+                    (i == len(args) - 1 or args[i + 1].startswith("--"))):
+                test_stdin = True
 
     # If --clear-logs the log files are cleared
     if "--clear-logs" in args:
@@ -704,6 +715,13 @@ def main(args=sys.argv[1:]):
         parser.error("Evaluation for multi-label models is not "
                      "available still.")
 
+    if train_stdin and command_args.multi_label:
+        parser.error("Reading multi-label training sets from stream "
+                     "is not yet available.")
+
+    if test_stdin and command_args.resume:
+        parser.error("Can't resume when using stream reading test sets.")
+
     default_output = ('evaluation' if command_args.evaluate
                       else 'predictions.csv')
     if command_args.resume:
@@ -713,9 +731,14 @@ def main(args=sys.argv[1:]):
         args = shlex.split(command)[1:]
         try:
             position = args.index("--train")
-            if (position == (len(args) - 1) or
-                    args[position + 1].startswith("--")):
-                train_stdin = True
+            train_stdin = (position == (len(args) - 1) or
+                           args[position + 1].startswith("--"))
+        except ValueError:
+            pass
+        try:
+            position = args.index("--test")
+            test_stdin = (position == (len(args) - 1) or
+                          args[position + 1].startswith("--"))
         except ValueError:
             pass
         output_dir = u.get_log_reversed(DIRS_LOG,
@@ -775,7 +798,12 @@ def main(args=sys.argv[1:]):
         command_args.debug = True
 
     if train_stdin:
+        if test_stdin:
+            sys.exit("The standard input can't be used both for training and"
+                     " testing. Choose one of them")
         command_args.training_set = StringIO.StringIO(sys.stdin.read())
+    elif test_stdin:
+        command_args.test_set = StringIO.StringIO(sys.stdin.read())
 
     api_command_args = {
         'username': command_args.username,
@@ -847,13 +875,13 @@ def main(args=sys.argv[1:]):
 
     # Parses dataset fields if provided.
     if command_args.dataset_fields:
-        dataset_fields_arg = map(lambda x: x.strip(),
+        dataset_fields_arg = map(str.strip,
                                  command_args.dataset_fields.split(','))
         output_args.update(dataset_fields=dataset_fields_arg)
 
     # Parses model input fields if provided.
     if command_args.model_fields:
-        model_fields_arg = map(lambda x: x.strip(),
+        model_fields_arg = map(str.strip,
                                command_args.model_fields.split(','))
         output_args.update(model_fields=model_fields_arg)
 
@@ -913,7 +941,7 @@ def main(args=sys.argv[1:]):
 
     # Reads votes files in the provided directories.
     if command_args.votes_dirs:
-        dirs = map(lambda x: x.strip(), command_args.votes_dirs.split(','))
+        dirs = map(str.strip, command_args.votes_dirs.split(','))
         votes_path = os.path.dirname(command_args.predictions)
         votes_files = u.read_votes_files(dirs, votes_path)
         output_args.update(votes_files=votes_files)
@@ -922,6 +950,11 @@ def main(args=sys.argv[1:]):
     if command_args.fields_map:
         fields_map_arg = u.read_fields_map(command_args.fields_map)
         output_args.update(fields_map=fields_map_arg)
+
+    # Old value for --prediction-info='full data' maps to 'full'
+    if command_args.prediction_info == 'full data':
+        print "WARNING: 'full data' is a deprecated value. Use 'full' instead"
+        command_args.prediction_info = FULL_FORMAT
 
     # Parses resources ids if provided.
     if command_args.delete:
@@ -935,7 +968,7 @@ def main(args=sys.argv[1:]):
                       console=command_args.verbosity)
         delete_list = []
         if command_args.delete_list:
-            delete_list = map(lambda x: x.strip(),
+            delete_list = map(str.strip,
                               command_args.delete_list.split(','))
         if command_args.delete_file:
             if not os.path.exists(command_args.delete_file):